requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
attrs==25.4.0
autocommand==2.2.2
Brlapi==0.8.7
btrfsutil==6.19
cffi==2.0.0
chardet==6.0.0.post1
click==8.3.1
cryptography==46.0.5
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
evdev==1.9.3
filelock==3.24.0
Glances==4.5.1
gps==3.27.5
hyprshade==4.0.1
i3ipc==2.2.1
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
input-remapper==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.6
keyutils==0.6
lit==21.1.8.dev0
louis==3.36.0
Mako==1.3.10.dev0
Markdown==3.10.2
MarkupSafe==3.0.3
more-itertools==10.8.0
netsnmp-python==1.0a1
nftables==0.1
numpy==2.4.3
nwg-displays==0.3.28
orjson==3.11.7
packaging==26.0
pillow==12.1.1
pipx==1.8.0
pkg_resources==81.0.0
platformdirs==4.9.4
psutil==7.2.2
pycairo==1.29.0
pycparser==2.23
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pydbus==0.6.0
PyGObject==3.54.5
PyQt5==5.15.11
PyQt5_sip==12.18.0
python-xlib==0.33
pywalfox==2.7.4
pyxdg==0.28
PyYAML==6.0.3
QTermWidget==2.3.0
Reflector==2023.6.28.0.36.1
reportlab==4.2.2
screeninfo==0.8.1
setproctitle==1.3.7
setuptools==82.0.0
shtab==1.8.0
six==1.17.0
smbus==1.1
TBB==0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
udiskie==2.6.1
ufw==0.36.2
userpath==1.9.2
uv==0.10.9
waypaper==2.7
wheel==0.46.3
python-decouple==3.8
requests==2.32.5
